pyproject.toml
[project]
name = "diff-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"fastapi>=0.116.0",
	"numpy>=1.26.0",
	"pillow>=10.0.0",
	"python-multipart>=0.0.9",
	"segmentation-models-pytorch>=0.4.0",
	"torch>=2.4.0",
	"torchvision>=0.19.0",
	"tqdm>=4.66.0",
	"uvicorn>=0.35.0",
]
